requirements.txt
asgiref==3.2.10
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djoser==2.1.0
drf-extra-fields==3.4.1
djangorestframework-simplejwt==4.8.0
idna==3.3
importlib-metadata==1.7.0
flake8==5.0.4
gunicorn==20.0.4
Pillow==9.2.0
PyJWT==2.4.0
pytz==2022.2
psycopg2-binary==2.8.6
python-dotenv==0.20.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.11
zipp==3.8.1
requests==2.28.1
requests-oauthlib==1.3.1
sorl-thumbnail==12.9.0